% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_message.R
\docType{data}
\name{.x}
\alias{.x}
\title{Startup message when package is attached.}
\format{An object of class \code{character} of length 5.}
\usage{
.x
}
\description{
Startup message when package is attached.
}
\keyword{datasets}
