% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_filters.R
\name{make_mke_schools}
\alias{make_mke_schools}
\title{Make common filterd dataframes from \code{wisconsink12} data.}
\usage{
make_mke_schools()
}
\description{
\code{make_mke_schools} filters the \code{schools} data for
  schools in the city of Milwaukee.
}
