% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_card_colors.R
\docType{data}
\name{significantly_exceeds_expectations_color}
\alias{significantly_exceeds_expectations_color}
\title{Report Card rating color codes}
\format{
An object of class \code{character} of length 1.
}
\usage{
significantly_exceeds_expectations_color
}
\description{
Stores the hex codes for colors associated with Wisconsin Report Card Ratings.
}
\keyword{datasets}
