% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisconsink12.R
\docType{package}
\name{wisconsink12}
\alias{wisconsink12}
\title{wisconsink12: A package for easy access to Wisconsin K12 school data}
\description{
The wisconsink12 package provides publicly-available data on public and private
Wisconsin K12 schools in a coherent relational database structure.
}
