% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_filters.R
\name{make_mke_schools}
\alias{make_mke_schools}
\title{Make common filterd dataframes from \code{wisconsink12} data.}
\usage{
make_mke_schools()
}
\arguments{
\item{private_type}{Select which type of report card to be included for choice schools.
Options are 'choice' for the 'Private - Choice Students' and 'all' for 'Private - All Students'
report card types.}

\item{exclude_milwaukee}{Logical. If TRUE (default value), Milwaukee schools will be
excluded from \code{wi_rc}.}
}
\description{
\code{make_mke_schools} Filters the \code{schools} data for
  schools in the city of Milwaukee.
\code{make_mke_rc} Filters the \code{report_cards} data for
  schools in the city of Milwaukee.
\code{make_wi_rc} Filters the \code{report_cards} data for
  school in Wisconsin, either including or excluding Milwaukee.
}
