% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_proficiency.R
\name{make_proficiency}
\alias{make_proficiency}
\title{Get school-level proficiency data}
\usage{
make_proficiency()
}
\description{
Creates a dataframe of school-level proficiency for all tested grades.
}
